SUBMARINE-257. Submarine web user manager page with angular
### What is this PR for?
Submarine web, add user manager page with angular

### What type of PR is it?
[Feature]

### What is the Jira issue?
https://issues.apache.org/jira/browse/SUBMARINE-257

### How should this be tested?
https://travis-ci.org/lleohao/hadoop-submarine/builds/609567077

### Screenshots (if appropriate)

1. user list
![image](https://user-images.githubusercontent.com/3677382/68528827-cc5b3900-0332-11ea-9094-081ec7e12c8c.png)

2. add user
![image](https://user-images.githubusercontent.com/3677382/68528836-e1d06300-0332-11ea-93f5-61ee42a2a53c.png)

3. view user detail

![image](https://user-images.githubusercontent.com/3677382/68528857-088e9980-0333-11ea-9747-fc992451983e.png)

4. edit user
![image](https://user-images.githubusercontent.com/3677382/68528859-0debe400-0333-11ea-8d23-25e0535a2277.png)

5. change user password
![image](https://user-images.githubusercontent.com/3677382/68528851-fca2d780-0332-11ea-82fc-b2d42e8f5eb9.png)

6. delete user
![image](https://user-images.githubusercontent.com/3677382/68528848-f7de2380-0332-11ea-939c-b4d9a751158e.png)

7. Log out
![image](https://user-images.githubusercontent.com/3677382/68528854-02002200-0333-11ea-9269-bdfea5664c9a.png)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: lleohao <lleohao@hotmail.com>

Closes #89 from lleohao/SUBMARINE-257 and squashes the following commits:

66e7c44 [lleohao] Merge branch 'master' into SUBMARINE-257
d6ddf35 [lleohao] [SUBMARINE-257] ci: fix build ng error
af4dff5 [lleohao] [SUBMARINE-257] test: fix e2e test case
79663bb [lleohao] [SUBMARINE-257] ci: add unit test
b977f89 [lleohao] [SUBMARINE-257] feat: add manager user feature
3d7055e [lleohao] [SUBMARINE-257] feat: add page header and logout feature
028b399 [lleohao] [SUBMARINE-257] feat: add sidebar and logo
20b96b6 [lleohao] [SUBMARINE-257] feat: init workbench module

+++ submarine-workbench/workbench-web-ng/src/app/services/auth.service.ts
@@ -19,12 +19,12 @@
 
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Rest, User } from '@submarine/interfaces';
-import { BaseApiService } from '@submarine/services/base-api.service';
-import { LocalStorageService } from '@submarine/services/local-storage.service';
+import { Rest, SysUser } from '@submarine/interfaces';
 import * as md5 from 'md5';
-import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { of, Observable } from 'rxjs';
+import { map, switchMap } from 'rxjs/operators';
+import { BaseApiService } from './base-api.service';
+import { LocalStorageService } from './local-storage.service';
 
 @Injectable({
   providedIn: 'root'
@@ -45,26 +45,36 @@ export class AuthService {
     this.isLoggedIn = !!authToken;
   }
 
-  login(userForm: { userName: string; password: string }): Observable<boolean> {
-    return this.httpClient
-      .post<Rest<User>>(this.baseApi.getRestApi('/auth/login'), {
-        username: userForm.userName,
-        password: md5(userForm.password)
-      })
-      .pipe(
-        map(res => {
-          if (res.success) {
-            this.isLoggedIn = true;
-            this.localStorageService.set(this.authTokenKey, res.result.token);
-          }
+  login(userForm: { userName: string; password: string }): Observable<SysUser> {
+    const apiUrl = this.baseApi.getRestApi('/auth/login');
+    const params = {
+      username: userForm.userName,
+      password: md5(userForm.password)
+    };
 
-          return res.success;
-        })
-      );
+    return this.httpClient.post<Rest<SysUser>>(apiUrl, params).pipe(
+      switchMap(res => {
+        if (res.success) {
+          this.isLoggedIn = true;
+          this.localStorageService.set(this.authTokenKey, res.result.token);
+          return of(res.result);
+        } else {
+          throw this.baseApi.createRequestError(res.message, res.code, apiUrl, 'post', params);
+        }
+      })
+    );
   }
 
-  logout(): void {
-    this.isLoggedIn = false;
-    this.localStorageService.remove(this.authTokenKey);
+  logout() {
+    return this.httpClient.post<Rest<boolean>>(this.baseApi.getRestApi('/auth/logout'), {}).pipe(
+      map(res => {
+        if (res.result) {
+          this.isLoggedIn = false;
+          this.localStorageService.remove(this.authTokenKey);
+        }
+
+        return res.result;
+      })
+    );
   }
 }
